Convert ExtensionsFragment.kt to Composable UI BatterySyncViewModel.kt and LockPhoneViewModel.kt have been consolidated into ExtensionsViewModel.kt
