+++ wearos/src/main/java/com/boswelja/devicemanager/phonelocking/ui/LockPhoneFragment.kt
@@ -1,9 +1,5 @@
 package com.boswelja.devicemanager.phonelocking.ui
 
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -17,43 +13,11 @@ import androidx.compose.material.Text
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.outlined.PhonelinkLock
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.livedata.observeAsState
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.ComposeView
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
-import androidx.fragment.app.Fragment
-import androidx.lifecycle.viewmodel.compose.viewModel
 import com.boswelja.devicemanager.R
-import com.boswelja.devicemanager.common.AppTheme
-
-class LockPhoneFragment : Fragment() {
-
-    @ExperimentalMaterialApi
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        return ComposeView(requireContext()).apply {
-            setContent {
-                AppTheme {
-                    val viewModel: LockPhoneViewModel = viewModel()
-                    val phoneLockingEnabled by viewModel.phoneLockingEnabled.observeAsState()
-                    val phoneName by viewModel.phoneName.observeAsState()
-                    PhoneLockingScreen(
-                        phoneLockingEnabled == true,
-                        phoneName ?: stringResource(R.string.default_phone_name)
-                    ) {
-                        viewModel.requestLockPhone()
-                    }
-                }
-            }
-        }
-    }
-}
 
 @ExperimentalMaterialApi
 @Composable
@@ -62,7 +26,7 @@ fun PhoneLockingScreen(
     phoneName: String,
     onClick: () -> Unit
 ) {
-    if (phoneLockingEnabled == true) {
+    if (phoneLockingEnabled) {
         PhoneLockingEnabled(
             phoneName = phoneName,
             onClick = onClick
